backend/requirements.txt
# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic[email]==2.6.1
pydantic-settings==2.1.0

# Database & ORM
sqlalchemy==2.0.25
asyncpg==0.29.0
pgvector==0.2.5
alembic==1.18.4

# We will need these later for Auth and Tasks, adding them now so the file is complete
python-jose[cryptography]==3.3.0
argon2-cffi==25.1.0
python-multipart==0.0.9
celery==5.3.6
redis==5.0.1
flower==2.0.1
python-json-logger==2.0.7

pytest==9.0.2
pytest-asyncio==1.3.0
httpx==0.28.1
pytest-cov==7.0.0